banco de dados alterado

+++ VaralChic/src/VaralChic/SQL/mydbvaralchic.sql
@@ -33,7 +33,7 @@ codigo_venda int not null primary key auto_increment,
 codigo_cliente int not null,
 codigo_produto int not null,
 data_venda date not null,
-valor decimal(10, 2) not null,
+valor float not null,
 metodo_pagamento varchar(100) not null,
 foreign key (codigo_cliente) references cliente (codigo_cliente),
 foreign key (codigo_produto) references produto (codigo_produto)
